Add MongoDb as a source

+++ Movies_API/Model/Movie.cs
@@ -11,13 +11,13 @@ namespace Movies_API.Model
         public string? Title { get; set; }
         public string? Description { get; set; }
         public string? ReleaseDate { get; set; }
-        public float VoteAverage { get; set; }
+        public double VoteAverage { get; set; }
         public int VoteCount { get; set; }
 
         public int RunTime { get; set; }
         public int Revenue { get; set; }
         public int Budget { get; set; }
-        public float Popularity { get; set; }
+        public double Popularity { get; set; }
 
         public string? PosterUrl { get; set; }
